+++ jkniv-jaas-glassfish/src/main/java/net/sf/jkniv/jaas/gf/HybridRealm.java
@@ -113,13 +113,13 @@ public class HybridRealm extends AppservRealm
         boolean authJdbc = false;
         boolean authCouch = false;
         
-        //        LOG.log(Level.FINEST,"LEVEL FINEST");
-        //        LOG.log(Level.FINER,"LEVEL FINER");
-        //        LOG.log(Level.FINE,"LEVEL FINE");
-        //        LOG.log(Level.INFO,"LEVEL INFO");
-        //        LOG.log(Level.CONFIG,"LEVEL CONFIG");
-        //        LOG.log(Level.WARNING,"LEVEL WARNING");
-        //        LOG.log(Level.SEVERE,"LEVEL SEVERE");
+        //LOG.log(Level.FINEST,"LEVEL FINEST");
+        //LOG.log(Level.FINER,"LEVEL FINER");
+        //LOG.log(Level.FINE,"LEVEL FINE");
+        //LOG.log(Level.INFO,"LEVEL INFO");
+        //LOG.log(Level.CONFIG,"LEVEL CONFIG");
+        //LOG.log(Level.WARNING,"LEVEL WARNING");
+        //LOG.log(Level.SEVERE,"LEVEL SEVERE");
         
         LOG.finest(
                 I18nManager.getString("hybrid.realm.infoauth", 
@@ -140,10 +140,10 @@ public class HybridRealm extends AppservRealm
         if (supportsAuthLdap)
             authLdap = ldapAdapter.authenticate(username, password, supportsAuthoLdap);
         
-        if (supportsAuthJdbc)// && !authLdap)
+        if (supportsAuthJdbc && !ldapAdapter.isRequisite(username))
             authJdbc = jdbcAdapter.authenticate(username, password);
         
-        if (supportsAuthCouch)// && !authLdap && !authJdbc)
+        if (supportsAuthCouch && !ldapAdapter.isRequisite(username))
             authCouch = couchDbAdapter.authenticate(username, password);
 
         if (!authLdap && !authJdbc && !authCouch)